Rename tracker files, more cleanup
