+++ rt_gene/src/rt_gene/__init__.py
@@ -2,8 +2,8 @@ from . import extract_landmarks_method
 from . import gaze_tools
 from . import kalman_stabilizer
 from . import subject_ros_bridge
-from . import GenericTracker
-from . import FaceEncodingTracker
-from . import SequentialTracker
+from . import tracker_generic
+from . import tracker_face_encoding
+from . import tracker_sequential
 
 from . import ThreeDDFA
